+++ packages/cosmwasm-stargate/src/cosmwasmclient.spec.ts
@@ -289,6 +289,21 @@ describe("CosmWasmClient", () => {
     });
   });
 
+  describe("getContractsByCreator", () => {
+    it("works", async () => {
+      pendingWithoutWasmd();
+      const client = await CosmWasmClient.connect(wasmd.endpoint);
+      const result = await client.getContractsByCreator(alice.address0);
+      const expectedAddresses = deployedHackatom.instances.map((info) => info.address);
+
+      // Test first 3 instances we get from scripts/wasmd/init.sh. There may me more than that in the result.
+      expect(result.length).toBeGreaterThanOrEqual(3);
+      expect(result[0]).toEqual(expectedAddresses[0]);
+      expect(result[1]).toEqual(expectedAddresses[1]);
+      expect(result[2]).toEqual(expectedAddresses[2]);
+    });
+  });
+
   describe("getContract", () => {
     it("works for instance without admin", async () => {
       pendingWithoutWasmd();